HWKMETRICS-421 Hawkular Metrics versioned support
Check version on graph panel initialization and then decide if we can
use the new POST method endpoints or fallback to a legacy call.

Legacy call uses deprecated data GET method endpoints, and applies rate
calculation on the client side. This is because old versions don't have
the rate endpoint for gauges.